Add Status & Activity Model
