+++ models/consts.py
@@ -1,5 +1,8 @@
 K_POST = 1001
 K_COMMENT = 1002
+K_STATUS = 1003
+K_ACTIVITY = 1004
+
 
 ONE_MINUTE = 60
 ONE_HOUR = ONE_MINUTE * 60
